Add files via upload
I added in a PlayerController class. Now we can map the different actions to different keys by editing the PlayerController. I also added in the button to switch to a second course to demonstrate the scene-changing mechanic. It is an instant switch.

But I still haven't figured out how to get the texture to carry over to the sprite properly. I appreciate the help!