Add files via upload
I added in a new flying enemy. The "Funbat" will fly erratically towards the player. It's also smaller than normal. These two traits make it hard to hit and really put the "fun" in "funbat." This is easily the craziest enemy type yet. Just a handful of them in a particular area is enough to cause a lot of trouble. Truly some serious bat-titude happening here. Oh dear