Add files via upload
I updated the Player class and the HealthBar class.

The HealthBar is now a Phaser.GameObjects.Graphics object instead of a Sprite, and I introduced more named variables to make it easier to adjust its position and appearance.
I also updated the Player class so that the HealthBar is loaded inside the Player's constructor and updated through the Player's update method. That way, the Player can pass the maxHP and current HP values directly to the HealthBar.