Add files via upload
I finished implementing every key animation event except for the death animation.
The spritesheet in the player folder has the "first draft" I created for all the important sprites needed.
For the most part, I would like the character's body and feet to stay in roughly the same spots for the final sprite sheet. For the walking/running animations, the walk cycles can definitely be improved. I'm only concerned that the body stays near the center for those animations.