Add files via upload
The new sprite sheets in here are from Nicoletta. I have them here in case Nabeeha's sprite sheet isn't done on time. These aren't done either, so it's a matter of which one is done first. The game is still set up to use the old sprite sheet, but that is easily changed